TT-6036 Handle many back translations
- TT04046 fix issue with sorting play item
- also fixes TT-6007
