+++ src/control/StyledBox.tsx
@@ -1,0 +1,13 @@
+import { Box, BoxProps, styled } from '@mui/material';
+
+interface StyledBoxProps extends BoxProps {
+  width?: number;
+}
+export const StyledBox = styled(Box, {
+  shouldForwardProp: (prop) => prop !== 'width',
+})<StyledBoxProps>(({ width, theme }) => ({
+  padding: `${theme.spacing(2)}px ${theme.spacing(6)}px`,
+  display: 'flex',
+  flexDirection: 'row',
+  width: `${width}px`,
+}));
